move magics to new jupyter-ai-magics project
