+++ lerna.json
@@ -2,5 +2,6 @@
   "$schema": "node_modules/lerna/schemas/lerna-schema.json",
   "useWorkspaces": true,
   "version": "0.4.0",
-  "npmClient": "yarn"
+  "npmClient": "yarn",
+  "useNx": true
 }
