+++ package.json
@@ -23,20 +23,21 @@
     "packages/*"
   ],
   "scripts": {
-    "setup:dev": "lerna run setup:dev --stream",
     "build": "lerna run build --stream",
     "build:core": "lerna run build --stream --scope \"@jupyter-ai/core\"",
     "build:prod": "lerna run build:prod --stream",
     "clean": "lerna run clean",
     "clean:all": "lerna run clean:all",
     "dev": "jupyter lab --config playground/config.py",
+    "dev-install": "lerna run dev-install --stream",
     "lint": "jlpm && lerna run prettier && lerna run eslint",
     "lint:check": "lerna run prettier:check && lerna run eslint:check",
     "watch": "lerna run watch --parallel --stream"
   },
   "devDependencies": {
     "@jupyterlab/builder": "^3.1.0",
-    "lerna": "^6.4.1"
+    "lerna": "^6.4.1",
+    "nx": "^15.9.2"
   },
   "nx": {
     "includedScripts": []
