+++ packages/jupyter-ai-magics/README.md
@@ -1,0 +1,4 @@
+# @jupyter-ai/magics
+
+This is a Python library `jupyter_ai_magics` that exposes the model providers
+and defines the IPython magics.
