+++ packages/jupyter-ai-magics/jupyter_ai_magics/providers.py
