+++ packages/jupyter-ai-magics/jupyter_ai_magics/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Python unit tests for jupyter_ai_magics."""
