+++ packages/jupyter-ai-magics/setup.py
@@ -1,0 +1,1 @@
+__import__('setuptools').setup()
