+++ packages/jupyter-ai/project.json
@@ -1,0 +1,4 @@
+{
+  "name": "@jupyter-ai/core",
+  "implicitDependencies": ["@jupyter-ai/magics"]
+}
