+++ packages/jupyter-ai/pyproject.toml
@@ -28,7 +28,8 @@ dependencies = [
     "openai~=0.26",
     "aiosqlite~=0.18",
     "importlib_metadata~=5.2.0",
-    "langchain~=0.0.115"
+    "langchain~=0.0.115",
+    "jupyter_ai_magics"
 ]
 
 dynamic = ["version", "description", "authors", "urls", "keywords"]
@@ -38,15 +39,6 @@ gpt3 = "jupyter_ai:GPT3ModelEngine"
 
 [project.entry-points."jupyter_ai.default_tasks"]
 core_default_tasks = "jupyter_ai:tasks"
-
-[project.entry-points."jupyter_ai.model_providers"]
-ai21 = "jupyter_ai:AI21Provider"
-anthropic = "jupyter_ai:AnthropicProvider"
-cohere = "jupyter_ai:CohereProvider"
-huggingface_hub = "jupyter_ai:HfHubProvider"
-openai = "jupyter_ai:OpenAIProvider"
-openai-chat = "jupyter_ai:ChatOpenAIProvider"
-sagemaker-endpoint = "jupyter_ai:SmEndpointProvider"
 
 [project.optional-dependencies]
 test = [
